last touches

+++ routes/api.js
@@ -48,6 +48,7 @@ router.get("/all-books", (req, res) => {
   })
 })
 
+// Add a new book to the database
 router.post("/books", (req, res) => {
   if (req.isAuthenticated()) {
     // TODO: Make a request to the google books api
@@ -64,7 +65,8 @@ router.post("/books", (req, res) => {
           title: volumeInfo.title,
           thumbnail: volumeInfo.imageLinks.thumbnail,
           owner: req.user._id,
-          requestedBy: []
+          requestedBy: null,
+          requestApproved: false
         })
 
         book.save((err) => {
@@ -87,7 +89,7 @@ router.post("/books", (req, res) => {
 })
 
 router.delete("/books", (req, res) => {
-  Book.findOneAndRemove({ _id: req.query.id }, (err) => {
+  Book.findOneAndRemove({ owner: req.user._id, _id: req.query.id }, (err) => {
     if (err) {
       res.error(err)
     } else {
@@ -102,9 +104,9 @@ router.delete("/books", (req, res) => {
   })
 })
 
-// Get incoming requests
-router.get("/request", (req, res) => {
-  Book.find({ owner: req.user._id, requestedBy: { $gt: [] } }, (err, docs) => {
+// Get requests made by the user
+router.get("/out-requests", (req, res) => {
+  Book.find({ requestedBy: req.user._id }, (err, docs) => {
     if (err) {
       res.error(err)
     } else {
@@ -113,12 +115,49 @@ router.get("/request", (req, res) => {
   })
 })
 
-router.post("/request", (req, res) => {
-  Book.findByIdAndUpdate(req.body.id, { $push: { requestedBy: req.user._id } }, (err, docs) => {
+// Get incoming requests for the user's books
+router.get("/request", (req, res) => {
+  Book.find({ owner: req.user._id, requestedBy: { $ne: null } }, (err, docs) => {
     if (err) {
       res.error(err)
     } else {
       res.send(docs)
+    }
+  })
+})
+
+// FIXME: Actually not secure since logged-in users could overwrite other peoples requests
+router.post("/request", (req, res) => {
+  if (req.isAuthenticated()) {
+    Book.findByIdAndUpdate(req.body.id, { requestedBy: req.user._id }, (err, docs) => {
+      if (err) {
+        res.error(err)
+      } else {
+        res.send(docs)
+      }
+    })
+  }
+})
+
+// FIXME: Api is not secure
+router.put("/request", (req, res) => {
+  Book.findByIdAndUpdate(req.body.id, { requestApproved: true }, (err) => {
+    if (err) {
+      res.error(err)
+    } else {
+      res.send("The request was approved")
+    }
+  })
+})
+
+// Delete a request FIXME: Not secure
+router.delete("/request", (req, res) => {
+  Book.findByIdAndUpdate(req.query.id, { requestedBy: null }, (err) => {
+    if (err) {
+      res.error(err)
+    } else {
+      // TODO: Make proper response
+      res.send("The request was deleted")
     }
   })
 })
